Implement splitlist in New issue dialog #9820
Implement splitlist to new issue dialog, change type from publish to simple for ContentRowStatusProps.
