+++ modules/lib/src/main/resources/assets/js/v6/features/shared/lists/content-row/types.ts
@@ -28,7 +28,7 @@ export type ContentRowLabelProps = {
 };
 
 export type ContentRowStatusProps = {
-    variant?: 'diff' | 'publish' | 'none';
+    variant?: 'diff' | 'simple' | 'none';
     className?: string;
 };
