+++ modules/lib/src/main/resources/assets/js/v6/features/views/browse/layout/BrowseAppShell.tsx
@@ -4,6 +4,7 @@ import {CompareVersionsDialog} from '../../../shared/dialogs/compare-versions/Co
 import {DeleteDialog} from '../../../shared/dialogs/delete/DeleteDialog';
 import {DuplicateDialog} from '../../../shared/dialogs/duplicate/DuplicateDialog';
 import {IssueDialog} from '../../../shared/dialogs/issue/IssueDialog';
+import {NewIssueDialog} from '../../../shared/dialogs/new-issue/NewIssueDialog';
 import {MoveDialog} from '../../../shared/dialogs/move/MoveDialog';
 import {NewContentDialog} from '../../../shared/dialogs/new-content/NewContentDialog';
 import {PublishDialog} from '../../../shared/dialogs/publish/PublishDialog';
@@ -23,6 +24,7 @@ export const BrowseAppShell = (): ReactElement => {
             <DuplicateDialog />
             <MoveDialog />
             <IssueDialog />
+            <NewIssueDialog />
             <CompareVersionsDialog />
         </>
     );
