+++ modules/lib/src/main/resources/assets/js/v6/features/views/wizard/layout/WizardAppShell.tsx
@@ -3,6 +3,7 @@ import {CompareVersionsDialog} from '../../../shared/dialogs/compare-versions/Co
 import {DeleteDialog} from '../../../shared/dialogs/delete/DeleteDialog';
 import {DuplicateDialog} from '../../../shared/dialogs/duplicate/DuplicateDialog';
 import {IssueDialog} from '../../../shared/dialogs/issue/IssueDialog';
+import {NewIssueDialog} from '../../../shared/dialogs/new-issue/NewIssueDialog';
 import {MoveDialog} from '../../../shared/dialogs/move/MoveDialog';
 import {PublishDialog} from '../../../shared/dialogs/publish/PublishDialog';
 import {RequestPublishDialog} from '../../../shared/dialogs/requestPublish/RequestPublishDialog';
@@ -19,6 +20,7 @@ export const WizardAppShell = (): ReactElement => {
             <DuplicateDialog />
             <MoveDialog />
             <IssueDialog />
+            <NewIssueDialog />
             <CompareVersionsDialog />
         </>
     );
